requirement.txt
altgraph==0.17.4
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
macholib==1.16.3
packaging==24.2
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
requests==2.32.3
urllib3==2.3.0
